# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  vo.BasicResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      request_id:
        type: string
      timestamp:
        type: integer
      trace_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 一个基于 Gin 的论坛系统 API
  title: Tiny Forum API
  version: "1.0"
paths:
...
  /users/me/role:
    get:
      consumes:
      - application/json
      description: 从数据库查询当前登录用户的角色信息
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取当前用户角色
      tags:
      - 用户管理
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"